% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_preprocessing.R
\docType{data}
\name{static_data}
\alias{static_data}
\title{Static data used throughout the text preprocessing module}
\format{
An object of class \code{list} of length 5.
}
\usage{
static_data
}
\description{
Static data used throughout the text preprocessing module
}
\keyword{internal}
